pyproject.toml
[project]
name = "invoice-reconciliation-api"
version = "0.1.0"
description = "Multi-Tenant Invoice Reconciliation API"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "sqlalchemy>=2.0.25",
    "strawberry-graphql[fastapi]>=0.219.0",
    "pydantic-settings>=2.1.0",
    "aiosqlite>=0.19.0",
    "httpx>=0.26.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
    "pytest-cov>=4.1.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
